requirements.txt
flask
openai
transformers
torch
sentence-transformers
scikit-learn
numpy
python-dotenv
jieba
tiktoken
pandas
pyarrow
